requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-httpx>=0.35.0
respx>=0.21.0
